use extern form of where (not an intern), as it has bug, which leads to performance degradation
problem is, that if we doesn't replace it with parameters (as we do it in id == "xyz")
the resulting query is "WHERE (a0.`id` = 1)" instead of WHERE (a0.`id` = ?)"
it results in much entries in prepared statements cache and in performance degradation.

the query in form of string parameters (Note id too as string: ~s(id == "#{id}")
are interpolated right and not used as literals in query. Later for integer
should be fixed too.

this fix should be seen, only quick fix temporar fix, until it is properly
fixed in ecdo